Separate out LightStepSpan from tracer.cpp.
