+++ c++11/CMakeLists.txt
@@ -75,7 +75,8 @@ set(LIGHTSTEP_SRCS src/utility.cpp
                    src/binary_carrier.cpp
                    src/grpc_transporter.cpp
                    src/rpc_recorder.cpp
-                   src/lightstep_span_context.cpp)
+                   src/lightstep_span_context.cpp
+                   src/lightstep_span.cpp)
 
 add_library(lightstep_tracer $<TARGET_OBJECTS:lightstep_protobuf>
                              ${LIGHTSTEP_SRCS})
