+++ c++11/src/tracer.cpp
@@ -9,6 +9,7 @@
 #include <random>
 #include <tuple>
 #include <vector>
+#include "lightstep_span.h"
 #include "lightstep_span_context.h"
 #include "propagation.h"
 #include "recorder.h"
@@ -18,242 +19,6 @@ using namespace opentracing;
 namespace lightstep {
 std::shared_ptr<opentracing::Tracer> MakeLightStepTracer(
     std::unique_ptr<Recorder>&& recorder);
-
-//------------------------------------------------------------------------------
-// set_span_reference
-//------------------------------------------------------------------------------
-static bool set_span_reference(
-    const std::pair<SpanReferenceType, const SpanContext*>& reference,
-    std::unordered_map<std::string, std::string>& baggage,
-    collector::Reference& collector_reference) {
-  collector_reference.Clear();
-  switch (reference.first) {
-    case SpanReferenceType::ChildOfRef:
-      collector_reference.set_relationship(collector::Reference::CHILD_OF);
-      break;
-    case SpanReferenceType::FollowsFromRef:
-      collector_reference.set_relationship(collector::Reference::FOLLOWS_FROM);
-      break;
-  }
-  if (reference.second == nullptr) {
-    std::cerr << "LightStep: passed in null span reference.\n";
-    return false;
-  }
-  auto referenced_context =
-      dynamic_cast<const LightStepSpanContext*>(reference.second);
-  if (referenced_context == nullptr) {
-    std::cerr << "LightStep: passed in span reference of unexpected type.\n";
-    return false;
-  }
-  collector_reference.mutable_span_context()->set_trace_id(
-      referenced_context->trace_id());
-  collector_reference.mutable_span_context()->set_span_id(
-      referenced_context->span_id());
-
-  referenced_context->ForeachBaggageItem(
-      [&baggage](const std::string& key, const std::string& value) {
-        baggage[key] = value;
-        return true;
-      });
-
-  return true;
-}
-
-//------------------------------------------------------------------------------
-// compute_start_timestamps
-//------------------------------------------------------------------------------
-static std::tuple<SystemTime, SteadyTime> compute_start_timestamps(
-    const SystemTime& start_system_timestamp,
-    const SteadyTime& start_steady_timestamp) {
-  // If neither the system nor steady timestamps are set, get the tme from the
-  // respective clocks; otherwise, use the set timestamp to initialize the
-  // other.
-  if (start_system_timestamp == SystemTime() &&
-      start_steady_timestamp == SteadyTime()) {
-    return {SystemClock::now(), SteadyClock::now()};
-  }
-  if (start_system_timestamp == SystemTime()) {
-    return {convert_time_point<SystemClock>(start_steady_timestamp),
-            start_steady_timestamp};
-  }
-  if (start_steady_timestamp == SteadyTime()) {
-    return {start_system_timestamp,
-            convert_time_point<SteadyClock>(start_system_timestamp)};
-  }
-  return {start_system_timestamp, start_steady_timestamp};
-}
-
-//------------------------------------------------------------------------------
-// LightStepSpan
-//------------------------------------------------------------------------------
-namespace {
-class LightStepSpan : public Span {
- public:
-  LightStepSpan(std::shared_ptr<const Tracer>&& tracer, Recorder& recorder,
-                string_view operation_name, const StartSpanOptions& options)
-      : tracer_(std::move(tracer)),
-        recorder_(recorder),
-        operation_name_(operation_name) {
-    // Set the start timestamps.
-    std::tie(start_timestamp_, start_steady_) = compute_start_timestamps(
-        options.start_system_timestamp, options.start_steady_timestamp);
-
-    // Set any span references.
-    std::unordered_map<std::string, std::string> baggage;
-    references_.reserve(options.references.size());
-    collector::Reference collector_reference;
-    for (auto& reference : options.references) {
-      if (!set_span_reference(reference, baggage, collector_reference)) {
-        continue;
-      }
-      references_.push_back(collector_reference);
-    }
-
-    // Set tags.
-    for (auto& tag : options.tags) {
-      tags_[tag.first] = tag.second;
-    }
-
-    // Set SpanContext.
-    auto trace_id = references_.empty()
-                        ? GenerateId()
-                        : references_[0].span_context().trace_id();
-    auto span_id = GenerateId();
-    span_context_ = LightStepSpanContext(trace_id, span_id, std::move(baggage));
-  }
-
-  LightStepSpan(const LightStepSpan&) = delete;
-  LightStepSpan(LightStepSpan&&) = delete;
-  LightStepSpan& operator=(const LightStepSpan&) = delete;
-  LightStepSpan& operator=(LightStepSpan&&) = delete;
-
-  ~LightStepSpan() override {
-    if (!is_finished_) {
-      Finish();
-    }
-  }
-
-  void FinishWithOptions(const FinishSpanOptions& options) noexcept override
-      try {
-    // Ensure the span is only finished once.
-    if (is_finished_.exchange(true)) {
-      return;
-    }
-
-    auto finish_timestamp = options.finish_steady_timestamp;
-    if (finish_timestamp == SteadyTime()) {
-      finish_timestamp = SteadyClock::now();
-    }
-
-    collector::Span span;
-
-    // Set timing information.
-    auto duration = finish_timestamp - start_steady_;
-    span.set_duration_micros(
-        std::chrono::duration_cast<std::chrono::microseconds>(duration)
-            .count());
-    *span.mutable_start_timestamp() = ToTimestamp(start_timestamp_);
-
-    // Set references.
-    auto references = span.mutable_references();
-    references->Reserve(static_cast<int>(references_.size()));
-    for (const auto& reference : references_) {
-      *references->Add() = reference;
-    }
-
-    // Set tags, logs, and operation name.
-    {
-      std::lock_guard<std::mutex> lock(mutex_);
-      span.set_operation_name(std::move(operation_name_));
-      auto tags = span.mutable_tags();
-      tags->Reserve(static_cast<int>(tags_.size()));
-      for (const auto& tag : tags_) {
-        *tags->Add() = to_key_value(tag.first, tag.second);
-      }
-      auto logs = span.mutable_logs();
-      for (auto& log : logs_) {
-        *logs->Add() = log;
-      }
-    }
-
-    // Set the span context.
-    auto span_context = span.mutable_span_context();
-    span_context->set_trace_id(span_context_.trace_id());
-    span_context->set_span_id(span_context_.span_id());
-    auto baggage = span_context->mutable_baggage();
-    span_context_.ForeachBaggageItem(
-        [baggage](const std::string& key, const std::string& value) {
-          using StringMap = google::protobuf::Map<std::string, std::string>;
-          baggage->insert(StringMap::value_type(key, value));
-          return true;
-        });
-
-    // Record the span
-    recorder_.RecordSpan(std::move(span));
-  } catch (const std::bad_alloc&) {
-    // Do nothing if memory allocation fails.
-  }
-
-  void SetOperationName(string_view name) noexcept override try {
-    std::lock_guard<std::mutex> l(mutex_);
-    operation_name_ = name;
-  } catch (const std::bad_alloc&) {
-    // Don't change operation name if memory can't be allocated for it.
-  }
-
-  void SetTag(string_view key, const Value& value) noexcept override try {
-    std::lock_guard<std::mutex> l(mutex_);
-    tags_[key] = value;
-  } catch (const std::bad_alloc&) {
-    // Don't add the tag if memory can't be allocated for it.
-  }
-
-  void SetBaggageItem(string_view restricted_key,
-                      string_view value) noexcept override {
-    span_context_.set_baggage_item(restricted_key, value);
-  }
-
-  std::string BaggageItem(string_view restricted_key) const noexcept override
-      try {
-    return span_context_.baggage_item(restricted_key);
-  } catch (const std::bad_alloc&) {
-    return {};
-  }
-
-  void Log(std::initializer_list<std::pair<string_view, Value>>
-               fields) noexcept override try {
-    auto timestamp = SystemClock::now();
-    collector::Log log;
-    *log.mutable_timestamp() = ToTimestamp(timestamp);
-    auto key_values = log.mutable_keyvalues();
-    for (const auto& field : fields) {
-      *key_values->Add() = to_key_value(field.first, field.second);
-    }
-    logs_.emplace_back(std::move(log));
-  } catch (const std::bad_alloc&) {
-    // Do nothing if memory can't be allocted for log records.
-  }
-
-  const SpanContext& context() const noexcept override { return span_context_; }
-  const Tracer& tracer() const noexcept override { return *tracer_; }
-
- private:
-  // Fields set in StartSpan() are not protected by a mutex.
-  std::shared_ptr<const Tracer> tracer_;
-  Recorder& recorder_;
-  std::vector<collector::Reference> references_;
-  SystemTime start_timestamp_;
-  SteadyTime start_steady_;
-  LightStepSpanContext span_context_;
-
-  // Mutex protects tags_ and operation_name_.
-  std::atomic<bool> is_finished_{false};
-  std::mutex mutex_;
-  std::string operation_name_;
-  std::unordered_map<std::string, Value> tags_;
-  std::vector<collector::Log> logs_;
-};
-}  // namespace
 
 //------------------------------------------------------------------------------
 // LightStepTracerImpl
