+++ c++11/src/utility.cpp
@@ -126,7 +126,7 @@ static std::string to_json(const Value& value) {
 }
 
 //------------------------------------------------------------------------------
-// to_key_value
+// ToKeyValue
 //------------------------------------------------------------------------------
 namespace {
 struct ValueVisitor {
@@ -160,7 +160,7 @@ struct ValueVisitor {
 };
 }  // anonymous namespace
 
-collector::KeyValue to_key_value(string_view key, const Value& value) {
+collector::KeyValue ToKeyValue(string_view key, const Value& value) {
   collector::KeyValue key_value;
   key_value.set_key(key);
   ValueVisitor value_visitor{key_value, value};
